Minor refactor of verifier
