+++ phase1-verifier/src/lib.rs
@@ -5,6 +5,7 @@ extern crate serde_derive;
 #[macro_use]
 extern crate thiserror;
 
+pub mod coordinator_requests;
 pub mod errors;
 pub mod utils;
 pub mod verifier;
