+++ phase1-verifier/src/verifier.rs
@@ -1,13 +1,15 @@
-use crate::{errors::VerifierError, utils::authenticate};
+use crate::{
+    errors::VerifierError,
+    utils::{create_parent_directory, write_to_file},
+};
 use phase1::helpers::CurveKind;
 use phase1_cli::transform_pok_and_correctness;
 use phase1_coordinator::{environment::Environment, phase1_chunked_parameters, Participant};
 use snarkos_toolkit::account::{Address, ViewKey};
 use zexe_algebra::{Bls12_377, BW6_761};
 
-use reqwest::Client;
-use std::{str::FromStr, thread::sleep, time::Duration};
-use tracing::{debug, error, info, trace};
+use std::{fs, str::FromStr, thread::sleep, time::Duration};
+use tracing::{error, info, trace};
 
 ///
 /// This lock response bundles the data required for the verifier
@@ -69,241 +71,6 @@ impl Verifier {
     }
 
     ///
-    /// Attempts to acquire the lock on a chunk.
-    ///
-    /// On success, this function returns the `LockResponse`.
-    ///
-    /// On failure, this function returns a `VerifierError`.
-    ///
-    pub async fn lock_chunk(&self) -> Result<LockResponse, VerifierError> {
-        let coordinator_api_url = &self.coordinator_api_url;
-        let method = "post";
-        let path = "/coordinator/verifier/lock";
-
-        let view_key = ViewKey::from_str(&self.view_key)?;
-
-        let signature_path = format!("/api{}", path);
-        let authentication = authenticate(&view_key, &method, &signature_path)?;
-
-        info!("Verifier attempting to lock a chunk");
-
-        match Client::new()
-            .post(&format!("{}{}", &coordinator_api_url, &path))
-            .header("Authorization", authentication.to_string())
-            .send()
-            .await
-        {
-            Ok(response) => {
-                if !response.status().is_success() {
-                    error!("Verifier failed to acquire a lock on a chunk");
-                    return Err(VerifierError::FailedLock);
-                }
-
-                // Parse the lock response
-                let json_response = response.json().await?;
-                let lock_response = serde_json::from_value::<LockResponse>(json_response)?;
-                debug!("Decoded verifier lock response: {:#?}", lock_response);
-
-                Ok(lock_response)
-            }
-            Err(_) => {
-                error!("Request ({}) to lock a chunk.", path);
-                return Err(VerifierError::FailedRequest(
-                    path.to_string(),
-                    coordinator_api_url.to_string(),
-                ));
-            }
-        }
-    }
-
-    ///
-    /// Attempts to run verification in the current round for a given `verified_locator`
-    ///
-    /// This assumes that a valid challenge file has already been uploaded to the
-    /// coordinator at the given `verified_locator`.
-    ///
-    /// On success, the coordinator returns an { "status": "ok" } response.
-    ///
-    /// On failure, this function returns a `VerifierError`.
-    ///
-    pub async fn verify_contribution(&self, verified_locator: &str) -> Result<String, VerifierError> {
-        let coordinator_api_url = &self.coordinator_api_url;
-        let method = "post";
-        let path = format!("/coordinator/verify/{}", verified_locator);
-
-        let view_key = ViewKey::from_str(&self.view_key)?;
-
-        info!(
-            "Verifier running verification of a response file at {} ",
-            verified_locator
-        );
-
-        let signature_path = format!("/api{}", path.replace("./", ""));
-        let authentication = authenticate(&view_key, &method, &signature_path)?;
-        match Client::new()
-            .post(&format!("{}{}", &coordinator_api_url, &path))
-            .header("Authorization", authentication.to_string())
-            .send()
-            .await
-        {
-            Ok(response) => {
-                if !response.status().is_success() {
-                    error!("Failed to verify the challenge {}", verified_locator);
-                    return Err(VerifierError::FailedVerification(verified_locator.to_string()));
-                }
-
-                Ok(response.text().await?)
-            }
-            Err(_) => {
-                error!("Request ({}) to verify a contribution failed.", path);
-                return Err(VerifierError::FailedRequest(
-                    path.to_string(),
-                    coordinator_api_url.to_string(),
-                ));
-            }
-        }
-    }
-
-    ///
-    /// Attempts to download the unverified response file from the coordinator at
-    /// a given `response_locator`
-    ///
-    /// On success, this function returns the full response file buffer.
-    ///
-    /// On failure, this function returns a `VerifierError`.
-    ///
-    pub async fn download_response_file(&self, response_locator: &str) -> Result<Vec<u8>, VerifierError> {
-        let coordinator_api_url = &self.coordinator_api_url;
-        let method = "get";
-        let path = format!("/coordinator/locator/{}", response_locator);
-
-        let view_key = ViewKey::from_str(&self.view_key)?;
-
-        info!("Verifier downloading a response file at {} ", response_locator);
-
-        let signature_path = format!("/api{}", path.replace("./", ""));
-        let authentication = authenticate(&view_key, &method, &signature_path)?;
-        match Client::new()
-            .get(&format!("{}{}", &coordinator_api_url, &path))
-            .header("Authorization", authentication.to_string())
-            .send()
-            .await
-        {
-            Ok(response) => {
-                if !response.status().is_success() {
-                    error!("Failed to download the response file {}", response_locator);
-                    return Err(VerifierError::FailedResponseDownload(response_locator.to_string()));
-                }
-
-                Ok(response.bytes().await?.to_vec())
-            }
-            Err(_) => {
-                error!("Request ({}) to download a response file failed.", path);
-                return Err(VerifierError::FailedRequest(
-                    path.to_string(),
-                    coordinator_api_url.to_string(),
-                ));
-            }
-        }
-    }
-
-    ///
-    /// Attempts to download the challenge file from the coordinator at
-    /// a given `challenge_locator`
-    ///
-    /// On success, this function returns the full challenge file buffer.
-    ///
-    /// On failure, this function returns a `VerifierError`.
-    ///
-    pub async fn download_challenge_file(&self, challenge_locator: &str) -> Result<Vec<u8>, VerifierError> {
-        let coordinator_api_url = &self.coordinator_api_url;
-        let method = "get";
-        let path = format!("/coordinator/locator/{}", challenge_locator);
-
-        let view_key = ViewKey::from_str(&self.view_key)?;
-
-        info!("Verifier downloading a challenge file at {} ", challenge_locator);
-
-        let signature_path = format!("/api{}", path.replace("./", ""));
-        let authentication = authenticate(&view_key, &method, &signature_path)?;
-        match Client::new()
-            .get(&format!("{}{}", &coordinator_api_url, &path))
-            .header("Authorization", authentication.to_string())
-            .send()
-            .await
-        {
-            Ok(response) => {
-                if !response.status().is_success() {
-                    error!("Failed to download the challenge file {}", challenge_locator);
-                    return Err(VerifierError::FailedChallengeDownload(challenge_locator.to_string()));
-                }
-
-                Ok(response.bytes().await?.to_vec())
-            }
-            Err(_) => {
-                error!("Request ({}) to download a challenge file failed.", path);
-                return Err(VerifierError::FailedRequest(
-                    path.to_string(),
-                    coordinator_api_url.to_string(),
-                ));
-            }
-        }
-    }
-
-    ///
-    /// Attempts to upload the next challenge locator file to the coordinator
-    ///
-    /// On success, this function returns the full challenge file buffer.
-    ///
-    /// On failure, this function returns a `VerifierError`.
-    ///
-    pub async fn upload_next_challenge_locator_file(
-        &self,
-        next_challenge_locator: &str,
-        next_challenge_file_bytes: Vec<u8>,
-    ) -> Result<String, VerifierError> {
-        let coordinator_api_url = &self.coordinator_api_url;
-        let method = "post";
-        let path = format!("/coordinator/verification/{}", next_challenge_locator);
-
-        let view_key = ViewKey::from_str(&self.view_key)?;
-
-        let signature_path = format!("/api{}", path.replace("./", ""));
-        let authentication = authenticate(&view_key, &method, &signature_path)?;
-
-        info!(
-            "Verifier uploading a response with size {} to {} ",
-            next_challenge_file_bytes.len(),
-            next_challenge_locator
-        );
-
-        match Client::new()
-            .post(&format!("{}{}", &coordinator_api_url, &path))
-            .header("Authorization", authentication.to_string())
-            .header("Content-Type", "application/octet-stream")
-            .body(next_challenge_file_bytes)
-            .send()
-            .await
-        {
-            Ok(response) => {
-                if !response.status().is_success() {
-                    error!("Failed to upload the new challenge file {}", next_challenge_locator);
-                    return Err(VerifierError::FailedChallengeUpload(next_challenge_locator.to_string()));
-                }
-
-                Ok(response.text().await?)
-            }
-            Err(_) => {
-                error!("Request ({}) to upload a new challenge file failed.", path);
-                return Err(VerifierError::FailedRequest(
-                    path.to_string(),
-                    coordinator_api_url.to_string(),
-                ));
-            }
-        }
-    }
-
-    ///
     /// Start the verifier loop. Polls the coordinator to lock and verify chunks.
     ///
     /// 1. Attempts to lock a chunk
@@ -322,33 +89,16 @@ impl Verifier {
         loop {
             // Attempt to lock a chunk
             if let Ok(lock_response) = self.lock_chunk().await {
-                info!("Verifier locked a chunk {} ", lock_response.chunk_id);
-
-                // Download the response file from the coordinator
-                let response_file = match self.download_response_file(&lock_response.response_locator).await {
-                    Ok(response) => {
-                        info!(
-                            "Verifier downloaded a response file {} of size {}",
-                            lock_response.response_locator,
-                            response.len()
-                        );
-                        response
-                    }
-                    Err(err) => {
-                        info!(
-                            "Verifier failed to download a response file {} (error: {})",
-                            lock_response.response_locator, err
-                        );
-                        continue;
-                    }
-                };
-
-                // Update the .verified to .unverified in the response locator
+                // Extract the lock response attributes
+                let chunk_id = lock_response.chunk_id;
                 let challenge_file_locator = lock_response.challenge_locator;
+                let response_locator = lock_response.response_locator;
+                let next_challenge_locator = lock_response.next_challenge_locator;
 
+                info!("Verifier locked a chunk {} ", chunk_id);
                 trace!("challenge_file_locator: {}", challenge_file_locator);
-                trace!("response_locator: {}", lock_response.response_locator);
-                trace!("next_challenge_locator: {}", lock_response.next_challenge_locator);
+                trace!("response_locator: {}", response_locator);
+                trace!("next_challenge_locator: {}", next_challenge_locator);
 
                 // Download the challenge file from the coordinator
                 let challenge_file = match self.download_challenge_file(&challenge_file_locator).await {
@@ -372,43 +122,45 @@ impl Verifier {
                 // Write the challenge and response files to disk
 
                 info!(
-                    "Writing the response file (size: {}) {} to disk",
-                    response_file.len(),
-                    &lock_response.response_locator
-                );
-                let response_locator_path = std::path::Path::new(&lock_response.response_locator);
-                let response_locator_parent = response_locator_path.parent().unwrap();
-                std::fs::create_dir_all(response_locator_parent).unwrap();
-                if let Err(err) = std::fs::write(response_locator_path, response_file) {
-                    error!(
-                        "Error writing response file to path {} (error: {})",
-                        &lock_response.response_locator, err
-                    );
-                }
-
-                info!(
                     "Writing the challenge file (size: {}) {} to disk",
                     challenge_file.len(),
                     &challenge_file_locator
                 );
-                let challenge_locator_path = std::path::Path::new(&challenge_file_locator);
-                let challenge_locator_path_parent = challenge_locator_path.parent().unwrap();
-                std::fs::create_dir_all(challenge_locator_path_parent).unwrap();
-                if let Err(err) = std::fs::write(challenge_locator_path, challenge_file) {
-                    error!(
-                        "Error writing challenge file to path {}  (error: {})",
-                        &challenge_file_locator, err
-                    );
-                }
+                write_to_file(&challenge_file_locator, challenge_file);
 
+                // Download the response file from the coordinator
+                let response_file = match self.download_response_file(&response_locator).await {
+                    Ok(response) => {
+                        info!(
+                            "Verifier downloaded a response file {} of size {}",
+                            response_locator,
+                            response.len()
+                        );
+                        response
+                    }
+                    Err(err) => {
+                        info!(
+                            "Verifier failed to download a response file {} (error: {})",
+                            response_locator, err
+                        );
+                        continue;
+                    }
+                };
+
+                // Write the response to a local file
+                info!(
+                    "Writing the response file (size: {}) {} to disk",
+                    response_file.len(),
+                    &response_locator
+                );
+                write_to_file(&response_locator, response_file);
+
+                // Create the parent directory for the `next_challenge_locator`
                 info!(
                     "Initializing the next challenge locator directory: {}",
-                    lock_response.next_challenge_locator
+                    next_challenge_locator
                 );
-
-                let next_challenge_locator_path = std::path::Path::new(&lock_response.next_challenge_locator);
-                let next_challenge_locator_path_parent = next_challenge_locator_path.parent().unwrap();
-                std::fs::create_dir_all(next_challenge_locator_path_parent).unwrap();
+                create_parent_directory(&next_challenge_locator);
 
                 // Run the verification
 
@@ -418,8 +170,6 @@ impl Verifier {
                 let compressed_challenge = self.environment.compressed_inputs();
                 let compressed_response = self.environment.compressed_outputs();
 
-                let chunk_id = lock_response.chunk_id;
-
                 info!("Running verification on chunk {}", chunk_id);
 
                 match curve {
@@ -427,42 +177,41 @@ impl Verifier {
                         compressed_challenge,
                         &challenge_file_locator,
                         compressed_response,
-                        &lock_response.response_locator,
+                        &response_locator,
                         compressed_challenge,
-                        &lock_response.next_challenge_locator,
+                        &next_challenge_locator,
                         &phase1_chunked_parameters!(Bls12_377, settings, chunk_id),
                     ),
                     CurveKind::BW6 => transform_pok_and_correctness(
                         compressed_challenge,
                         &challenge_file_locator,
                         compressed_response,
-                        &lock_response.response_locator,
+                        &response_locator,
                         compressed_challenge,
-                        &lock_response.next_challenge_locator,
+                        &next_challenge_locator,
                         &phase1_chunked_parameters!(BW6_761, settings, chunk_id),
                     ),
                 };
 
-                let next_challenge_file = match std::fs::read(&lock_response.next_challenge_locator) {
+                // Reads the file at the `next_challenge_locator` that was generated
+                // by the verification step
+                let next_challenge_file = match fs::read(&next_challenge_locator) {
                     Ok(file) => {
-                        info!(
-                            "Reading next challenge locator at {}",
-                            &lock_response.next_challenge_locator
-                        );
+                        info!("Reading next challenge locator at {}", &next_challenge_locator);
                         file
                     }
                     Err(err) => {
                         error!(
                             "Failed to open next_challenge_file at {} (error {})",
-                            &lock_response.next_challenge_locator, err
+                            &next_challenge_locator, err
                         );
                         continue;
                     }
                 };
 
-                // Upload the verified response file to `lock_response.response.locator`
+                // Upload the new challenge file to `next_challenge_locator`
                 match self
-                    .upload_next_challenge_locator_file(&lock_response.next_challenge_locator, next_challenge_file)
+                    .upload_next_challenge_locator_file(&next_challenge_locator, next_challenge_file)
                     .await
                 {
                     Ok(response) => {
@@ -474,22 +223,22 @@ impl Verifier {
                     Err(err) => {
                         error!(
                             "Failed to open upload next challenge {} to coordinator (error: {})",
-                            &lock_response.next_challenge_locator, err
+                            &next_challenge_locator, err
                         );
                         continue;
                     }
                 };
 
-                // Attempt to apply the uploaded response locator
-                match self.verify_contribution(&lock_response.next_challenge_locator).await {
+                // Attempt to perform the verification with the uploaded challenge file at `next_challenge_locator`
+                match self.verify_contribution(&next_challenge_locator).await {
                     Ok(response) => info!(
                         "Verifier verified the response file {}. Response {}",
-                        lock_response.next_challenge_locator, response
+                        next_challenge_locator, response
                     ),
                     Err(err) => {
                         error!(
                             "Failed to run verification on the next challenge {} (error: {})",
-                            &lock_response.next_challenge_locator, err
+                            &next_challenge_locator, err
                         );
                         continue;
                     }
